+++ MagazineLayout/Public/MagazineLayout.swift
@@ -1532,20 +1532,3 @@ private extension MagazineLayout {
   }
 
 }
-
-// MARK: - Signposting
-
-let signpostLog = OSLog(subsystem: "com.airbnb.MagazineLayout", category: "MagazineLayout")
-
-enum SignpostName {
-  static let collectionViewContentSize: StaticString = "MagazineLayout.collectionViewContentSize"
-  static let prepare: StaticString = "MagazineLayout.prepare"
-  static let prepareUpdateWidths: StaticString = "MagazineLayout.prepare.prepareUpdateWidths"
-  static let prepareUpdateLayoutMetrics: StaticString = "MagazineLayout.prepare.prepareUpdateLayoutMetrics"
-  static let prepareRecreateSectionModels: StaticString = "MagazineLayout.prepare.recreateSectionModels"
-  static let layoutAttributesForElementsInRect: StaticString = "MagazineLayout.layoutAttributesForElementsInRect"
-  static let prepareForCollectionViewUpdates: StaticString = "MagazineLayout.prepareForCollectionViewUpdates"
-  static let invalidateLayout: StaticString = "MagazineLayout.invalidateLayout"
-  static let preferredLayoutAttributesFittingCell: StaticString = "MagazineLayout.preferredLayoutAttributesFitting.cell"
-  static let preferredLayoutAttributesFittingReusableView: StaticString = "MagazineLayout.preferredLayoutAttributesFitting.reusableView"
-}
